+++ scripts/upfront/upfront-application.js
@@ -463,7 +463,6 @@ var PostLayoutEditor = new (LayoutEditorSubapplication.extend({
 	saveEdition: function(){
 		if(Application.current_subapplication != PostLayoutEditor)
 			return;
-
 		var me = this,
 			saveDialog = new Upfront.Views.Editor.SaveDialog({
 				question: Upfront.Settings.l10n.global.application.save_layout_pop,
@@ -1569,10 +1568,13 @@ var Application = new (Backbone.Router.extend({
 				_.each(elementStyles, function(style, name){
 					style = Upfront.Util.colors.convert_string_ufc_to_color(style);
 					Upfront.data.styles[elementType].push(name);
-					var styleNode = $(name);
+					var styleNode = $('#'+name);
 					if(!styleNode.length){
 						styleNode = $('<style id="' + name + '">' + style + '</style>');
 						$('body').append(styleNode);
+					}
+					else {
+						styleNode.html(style);
 					}
 				});
 			});