+++ CHANGELOG.md
@@ -2,6 +2,43 @@ Change Log
 ============
 
 
+0.2.7.1 - 2015-03-17
+-------------------------------------------------------------------------------
+- Fix: resizing handle hidden when editing elements in group.
+- Fix: hide settings button and resizable handle on group when editing elements.
+
+0.2.7 - 2015-03-17
+-------------------------------------------------------------------------------
+- Fix: custom 404 layout changes saving.
+- Fix: image embed in text/accordion editing.
+- Fix: remove gallery image rotate functionality.
+- Fix: clean up the passed popup classname parameter on close.
+- Fix: image warning popup styles.
+- Fix: skip prefixing the global CSS.
+- Fix: validate the selected image size argument for code element.
+- Fix: posts/pages popup bugging out if no author specified.
+- Fix: drag and drop issue on the last element
+
+0.2.6 - 2015-03-10
+-------------------------------------------------------------------------------
+- Fix: Fix image blocks UI when S3 plugins move images.
+- Fix: youtube element accept short ahare url format i.e. youtu.be
+- Fix: keep ratio behavior for full screen region
+- Fix: muted video background
+- Fix: theme colors in code element
+- Fix: post layout wont apply to all post types
+- Fix: accordion panel add button not showing
+- Fix: multiple spectrums open
+- Fix: disable alpha slider when theme color is chosen
+
+0.2.5 - 2015-03-05
+-------------------------------------------------------------------------------
+- Fix: gallery labels adding.
+- Fix: compensate for dead element double-click event.
+- Fix: like box fixed misalignment of thumbnails.
+- Fix: like box mapped return key to send action.
+- Fix: too many controlls after element group and chosen nag for sprites.
+
 0.2.4 - 2015-03-04
 -------------------------------------------------------------------------------
 - Fix: YouTube and tabs elements.